requirements.txt
flask==2.0.1
pypdf
langchain==0.1.0
langchain_community
langchain_pinecone
langchain_groq
pinecone-client
python-dotenv
gunicorn==20.1.0
sentence-transformers